Moved files into own directories. Updated paths for tests
